Alterando o blog

+++ src/main/java/org/Generation/BlogPessoal/Controller/PostagemController.java
@@ -1,12 +1,14 @@
-package org.Generation.BlogPessoal.Controller;
+package org.generation.BlogPessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
-import org.Generation.BlogPessoal.model.Postagem;
-import org.Generation.BlogPessoal.repository.PostagemRepository;
+import org.generation.BlogPessoal.model.Postagem;
+import org.generation.BlogPessoal.repository.PostagemRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -16,44 +18,55 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
-@RequestMapping("postagens")
+@RequestMapping("/postagem")
 @CrossOrigin(value = "*", allowedHeaders = "*")
 public class PostagemController {
-
+	
 	@Autowired
 	private PostagemRepository repository;
 
-	@GetMapping("/all")
-	public ResponseEntity<List<Postagem>> GetAll() {
+	@GetMapping
+	public ResponseEntity<List<Postagem>> getAll(){
 		return ResponseEntity.ok(repository.findAll());
 	}
 
 	@GetMapping("/{id}")
-	public ResponseEntity<Postagem> GetById(@PathVariable Long id) {
-		return repository.findById(id).map(resp -> ResponseEntity.ok(resp)).orElse(ResponseEntity.notFound().build());
+	public ResponseEntity<Postagem> getById(@PathVariable long id){
+		return repository.findById(id).map(resp -> ResponseEntity.ok(resp))
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
 
 	@GetMapping("/titulo/{titulo}")
-	public ResponseEntity<List<Postagem>> GetByTitulo(@PathVariable String titulo){
+	public ResponseEntity<List<Postagem>> getByTitle(@PathVariable String titulo){
 		return ResponseEntity.ok(repository.findAllByTituloContainingIgnoreCase(titulo));
 	}
 
-	@PostMapping("/post")
-	public ResponseEntity<Postagem> post(@Valid @RequestBody Postagem newUser) {
-		return ResponseEntity.status(201).body(repository.save(newUser));
-
+	@PostMapping
+	public ResponseEntity<Postagem> post(@Valid @RequestBody Postagem postagem){
+		return ResponseEntity.status(HttpStatus.CREATED)
+				.body(repository.save(postagem));
 	}
 
-	@PutMapping ("/put")
-	public ResponseEntity<Postagem> put(@Valid @RequestBody Postagem postagem) {
-		return ResponseEntity.status(201).body(repository.save(postagem));
+	@PutMapping
+	public ResponseEntity<Postagem> put(@Valid @RequestBody Postagem postagem){
+		return repository.findById(postagem.getId())
+				.map(resp -> ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem)))
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
 
-	@DeleteMapping("/delete/{id}")
+	@ResponseStatus(HttpStatus.NO_CONTENT)
+	@DeleteMapping("/{id}")
 	public void delete(@PathVariable long id) {
+		Optional<Postagem> post = repository.findById(id);
+
+		if(post.isEmpty())
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
 		repository.deleteById(id);
 	}
 }
